Started writing manual (WIP) and attempted fixed for hangs in grpc
it appears fetching the live logs occasionally hangs. Providing a deadline is supposed to address this
but doesn't. To work around, put the stub in its own thread, and if a call times out, throw an exception. A thread will get leaked each time this happens, but the watch process is fairly short lived so I don't see that as a problem.